upd: linker script

+++ hyperium/hyperiumtypes.hpp
@@ -11,7 +11,7 @@ using int16_t = short int;
 using uint8_t = unsigned char;
 using int8_t = char;
 
-using size_t = unsigned long long;
+using size_t = unsigned long ;
 using size_p = unsigned long long int;
 
 using Address = uint64_t;
